# openapi/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
...
paths:
...
  /api/v1/buildings/{buildingId}:
    get:
      tags:
      - buildings
      summary: Find building by ID
      operationId: findBuilding
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Building not found
          content: {}
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        "403":
          description: Access is forbidden to the resource
    put:
      tags:
      - buildings
      summary: Update existing building
      operationId: updateBuilding
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        description: Building object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Building'
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Building not found
          content: {}
        "200":
          description: Success
        "403":
          description: Access is forbidden to the resource
    delete:
      tags:
      - buildings
      summary: Delete existing building
      operationId: deleteBuilding
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Building not found
          content: {}
        "204":
          description: No content
        "403":
          description: Access is forbidden to the resource
...
  /api/v1/facilities:
    get:
      tags:
      - facilities
      summary: Get facility list
      operationId: findPage
      parameters:
      - name: offset
        in: query
        description: The offset of the first item in the collection to return
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: The maximum number of entries to return
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "403":
          description: Access is forbidden to the resources
    post:
      tags:
      - facilities
      summary: Add new facility
      operationId: createFacility
      requestBody:
        description: Facility object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Facility'
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
        "405":
          description: Invalid input
  /api/v1/buildings:
    get:
      tags:
      - buildings
      summary: Get building list
      operationId: findPage_1
      parameters:
      - name: offset
        in: query
        description: The offset of the first item in the collection to return
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: The maximum number of entries to return
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "403":
          description: Access is forbidden to the resources
    post:
      tags:
      - buildings
      summary: Add new building
      operationId: createBuilding
      requestBody:
        description: Building object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Building'
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
        "405":
          description: Invalid input
  /api/v1/buildings/{buildingId}/users/{userId}:
    post:
      tags:
      - buildings
      summary: Add new building user relationship
      operationId: addBuildingUser
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      - name: userId
        in: path
        description: ID of a user
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Building or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
    delete:
      tags:
      - buildings
      summary: Delete an existing building user relationship
      operationId: deleteBuildingUser
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      - name: userId
        in: path
        description: ID of a user
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Building or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "204":
          description: No content
  /api/v1/buildings/{buildingId}/rooms:
    get:
      tags:
      - rooms
      summary: Get room list
      operationId: findPage_2
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        description: The offset of the first item in the collection to return
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: The maximum number of entries to return
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Building not found
          content: {}
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "403":
          description: Access is forbidden to the resources
    post:
      tags:
      - rooms
      summary: Add new room
      operationId: createRoom
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        description: Room object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Building not found
          content: {}
        "201":
          description: Created
        "405":
          description: Invalid input
  /api/v1/buildings/{buildingId}/facilities/{facilityId}:
    post:
      tags:
      - buildings
      summary: Add new building facility relationship
      operationId: addBuildingFacility
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      - name: facilityId
        in: path
        description: ID of a facility
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Building or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
    delete:
      tags:
      - buildings
      summary: Delete an existing building facility relationship
      operationId: deleteBuildingFacility
      parameters:
      - name: buildingId
        in: path
        description: ID of a building
        required: true
        schema:
          type: integer
          format: int32
      - name: facilityId
        in: path
        description: ID of a facility
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "404":
          description: Building or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "204":
          description: No content
  /api/v1/users/{userId}/buildings:
    get:
      tags:
      - buildings
      summary: Get building list of a specific user
      operationId: findBuildingsOfUser
      parameters:
      - name: userId
        in: path
        description: ID of a user
        required: true
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        description: The offset of the first item in the collection to return
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: The maximum number of entries to return
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: User not found
          content: {}
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "403":
          description: Access is forbidden to the resources
  /api/v1/facilities/{facilityId}/buildings:
    get:
      tags:
      - buildings
      summary: Get building list of a specific facility
      operationId: findBuildingsOfFacility
      parameters:
      - name: facilityId
        in: path
        description: ID of a facility
        required: true
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        description: The offset of the first item in the collection to return
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        description: The maximum number of entries to return
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad request
          content: {}
        "404":
          description: Facility not found
          content: {}
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "403":
          description: Access is forbidden to the resources
...
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        description:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum:
          - AMMINISTRATORE
          - MANAGER
          - CLIENTE
          - RECEPTIONIST
        email:
          type: string
        telephoneNumber:
          type: string
    Facility:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        price:
          type: number
        flagLuxury:
          type: boolean
    Building:
      type: object
      properties:
        id:
          type: integer
          format: int32
        telephoneNumber:
          type: string
        email:
          type: string
        flagWiFi:
          type: boolean
        buildingType:
          type: string
          enum:
          - SPA
          - HOTEL
          - RESORT
          - BED_AND_BREAKFAST
          - AFFITTACAMERE
        street:
          type: string
        city:
          type: string
        country:
          type: string
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/Room'
        users:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/User'
        facilities:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Facility'
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int32
        size:
          type: number
        priceLowSeason:
          type: number
        priceMidSeason:
          type: number
        priceHighSeason:
          type: number
        flagBalcony:
          type: boolean
        flagShower:
          type: boolean
        flagWhirlpool:
          type: boolean
        category:
          type: string
          enum:
          - ECONOMY
          - FAMILY
          - JUNIOR_SUITE
          - DELUXE_SUITE
        singleBeds:
          type: number
        queenSizeBeds:
          type: number
        bathrooms:
          type: number
    Page:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            type: object
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...